trivial: split storage_interface::state_store::state
