+++ execution/executor/src/chunk_executor/transaction_chunk.rs
@@ -10,7 +10,7 @@ use aptos_executor_types::execution_output::ExecutionOutput;
 use aptos_experimental_runtimes::thread_manager::optimal_min_len;
 use aptos_metrics_core::TimerHelper;
 use aptos_storage_interface::state_store::{
-    state::LedgerState, state_view::cached_state_view::CachedStateView,
+    state::ledger_state::LedgerState, state_view::cached_state_view::CachedStateView,
 };
 use aptos_types::{
     block_executor::{
