+++ storage/aptosdb/src/state_store/state_merkle_batch_committer.rs
@@ -15,7 +15,9 @@ use aptos_jellyfish_merkle::node_type::NodeKey;
 use aptos_logger::{info, trace};
 use aptos_metrics_core::TimerHelper;
 use aptos_schemadb::batch::RawBatch;
-use aptos_storage_interface::state_store::{state::State, state_with_summary::StateWithSummary};
+use aptos_storage_interface::state_store::{
+    state::state::State, state_with_summary::StateWithSummary,
+};
 use std::sync::{mpsc::Receiver, Arc};
 
 pub struct StateMerkleBatch {
