+++ storage/storage-interface/src/state_store/state/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod ledger_state;
+#[allow(clippy::module_inception)]
+pub mod state;
