+++ storage/storage-interface/src/state_store/state/state.rs
@@ -5,23 +5,20 @@ use crate::{
     metrics::TIMER,
     state_store::{
         state_delta::StateDelta,
-        state_update_refs::{BatchedStateUpdateRefs, StateUpdateRefs},
-        state_view::cached_state_view::{CachedStateView, ShardedStateCache, StateCacheShard},
+        state_update_refs::BatchedStateUpdateRefs,
+        state_view::cached_state_view::{ShardedStateCache, StateCacheShard},
         versioned_state_value::{StateUpdate, StateUpdateRef},
         NUM_STATE_SHARDS,
     },
-    DbReader,
 };
-use anyhow::Result;
 use aptos_experimental_layered_map::{LayeredMap, MapLayer};
 use aptos_metrics_core::TimerHelper;
 use aptos_types::{
-    state_store::{state_key::StateKey, state_storage_usage::StateStorageUsage, StateViewId},
+    state_store::{state_key::StateKey, state_storage_usage::StateStorageUsage},
     transaction::Version,
 };
-use derive_more::Deref;
 use itertools::Itertools;
-use rayon::prelude::*;
+use rayon::iter::{IntoParallelIterator, ParallelIterator};
 use std::{collections::HashMap, sync::Arc};
 
 /// Represents the blockchain state at a given version.
@@ -186,101 +183,5 @@ impl State {
             }
         }
         (items_delta, bytes_delta)
-    }
-}
-
-/// At a given version, the state and the last checkpoint state at or before the version.
-#[derive(Clone, Debug, Deref)]
-pub struct LedgerState {
-    last_checkpoint: State,
-    #[deref]
-    latest: State,
-}
-
-impl LedgerState {
-    pub fn new(latest: State, last_checkpoint: State) -> Self {
-        assert!(latest.is_descendant_of(&latest));
-
-        Self {
-            latest,
-            last_checkpoint,
-        }
-    }
-
-    pub fn new_empty() -> Self {
-        let state = State::new_empty();
-        Self::new(state.clone(), state)
-    }
-
-    pub fn latest(&self) -> &State {
-        &self.latest
-    }
-
-    pub fn last_checkpoint(&self) -> &State {
-        &self.last_checkpoint
-    }
-
-    pub fn is_checkpoint(&self) -> bool {
-        self.latest.is_the_same(&self.last_checkpoint)
-    }
-
-    /// In the execution pipeline, at the time of state update, the reads during execution
-    /// have already been recorded.
-    pub fn update_with_memorized_reads(
-        &self,
-        persisted_snapshot: &State,
-        updates: &StateUpdateRefs,
-        reads: &ShardedStateCache,
-    ) -> LedgerState {
-        let _timer = TIMER.timer_with(&["ledger_state__update"]);
-
-        let last_checkpoint = if let Some(updates) = &updates.for_last_checkpoint {
-            self.latest().update(persisted_snapshot, updates, reads)
-        } else {
-            self.last_checkpoint.clone()
-        };
-
-        let base_of_latest = if updates.for_last_checkpoint.is_none() {
-            self.latest()
-        } else {
-            &last_checkpoint
-        };
-        let latest = if let Some(updates) = &updates.for_latest {
-            base_of_latest.update(persisted_snapshot, updates, reads)
-        } else {
-            base_of_latest.clone()
-        };
-
-        LedgerState::new(latest, last_checkpoint)
-    }
-
-    /// Old values of the updated keys are read from the DbReader at the version of the
-    /// `persisted_snapshot`.
-    pub fn update_with_db_reader(
-        &self,
-        persisted_snapshot: &State,
-        updates: &StateUpdateRefs,
-        reader: Arc<dyn DbReader>,
-    ) -> Result<(LedgerState, ShardedStateCache)> {
-        let state_view = CachedStateView::new_impl(
-            StateViewId::Miscellaneous,
-            reader,
-            persisted_snapshot.clone(),
-            self.latest().clone(),
-        );
-        state_view.prime_cache(updates)?;
-
-        let updated = self.update_with_memorized_reads(
-            persisted_snapshot,
-            updates,
-            state_view.memorized_reads(),
-        );
-        let state_reads = state_view.into_memorized_reads();
-        Ok((updated, state_reads))
-    }
-
-    pub fn is_the_same(&self, other: &Self) -> bool {
-        self.latest.is_the_same(&other.latest)
-            && self.last_checkpoint.is_the_same(&other.last_checkpoint)
     }
 }
